snowbridge_pallet_outbound_queue::pallet

Trait Config

pub trait Config: Config + Config<RuntimeEvent: From<Event<Self>>> {
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type Hashing: Hash<Output = H256>;
    type MessageQueue: EnqueueMessage<AggregateMessageOrigin>;
    type GasMeter: GasMeter;
    type Balance: Balance + From<u128>;
    type Decimals: Get<u8>;
    type MaxMessagePayloadSize: Get<u32>;
    type MaxMessagesPerBlock: Get<u32>;
    type Channels: Contains<ChannelId>;
    type PricingParameters: Get<PricingParameters<Self::Balance>>;
    type WeightToFee: WeightToFee<Balance = Self::Balance>;
    type WeightInfo: WeightInfo;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

type Hashing: Hash<Output = H256>

type MessageQueue: EnqueueMessage<AggregateMessageOrigin>

type GasMeter: GasMeter

Measures the maximum gas used to execute a command on Ethereum

type Balance: Balance + From<u128>

type Decimals: Get<u8>

Number of decimal places in native currency

type MaxMessagePayloadSize: Get<u32>

Max bytes in a message payload

type MaxMessagesPerBlock: Get<u32>

Max number of messages processed per block

type Channels: Contains<ChannelId>

Check whether a channel exists

type PricingParameters: Get<PricingParameters<Self::Balance>>

type WeightToFee: WeightToFee<Balance = Self::Balance>

Convert a weight value into a deductible fee based.

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.